refactor: move redirect to folder
